# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeScope API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history/{userId}:
    get:
      operationId: listAnalyses
      tags: [history]
      summary: Get analysis history for a user (last 50)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
  /history/{userId}/filter:
    post:
      operationId: filterAnalyses
      tags: [history]
      summary: Filter analysis history with criteria
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryFilter"
      responses:
        "200":
          description: Filtered history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"
  /history/{analysisId}/outcome:
    put:
      operationId: updateOutcome
      tags: [history]
      summary: Update the outcome of an analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutcomeUpdate"
      responses:
        "200":
          description: Updated analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /history/{analysisId}:
    get:
      operationId: getAnalysis
      tags: [history]
      summary: Get a single analysis by ID
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /propfirm/challenges/{userId}:
    get:
      operationId: getChallenges
      tags: [propfirm]
      summary: Get all prop firm challenges for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropChallenge"
  /propfirm/challenge:
    post:
      operationId: createChallenge
      tags: [propfirm]
      summary: Create a new prop firm challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeInput"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropChallenge"
  /propfirm/challenge/{id}:
    get:
      operationId: getChallenge
      tags: [propfirm]
      summary: Get a specific challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Challenge detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropChallenge"
    delete:
      operationId: deleteChallenge
      tags: [propfirm]
      summary: Delete a challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /propfirm/challenge/{id}/trade:
    put:
      operationId: logChallengeTrade
      tags: [propfirm]
      summary: Log a trade P&L against a challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeTradeInput"
      responses:
        "200":
          description: Updated challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropChallenge"
  /propfirm/challenge/{id}/simulate:
    post:
      operationId: simulateChallengeLoss
      tags: [propfirm]
      summary: Simulate a hypothetical loss on a challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
...
  /notes/{analysisId}:
    get:
      operationId: getTradeNote
      tags: [history]
      summary: Get trade journal note for an analysis
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeNote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertTradeNote
      tags: [history]
      summary: Create or update a trade journal note
      parameters:
        - name: analysisId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeNoteInput"
      responses:
        "200":
          description: Trade note saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeNote"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Confluence:
      type: object
      required: [trendAlignment, entryTiming, riskManagement, newsSafety, sessionQuality, reasoningQuality]
      properties:
        trendAlignment:
          type: integer
        entryTiming:
          type: integer
        riskManagement:
          type: integer
        newsSafety:
          type: integer
        sessionQuality:
          type: integer
        reasoningQuality:
          type: integer
        chartStructure:
          type: ["integer", "null"]
    NewsAlert:
      type: object
      required: [event, minutesAway, recommendation]
      properties:
        event:
          type: string
        minutesAway:
          type: integer
        recommendation:
          type: string
    Analysis:
      type: object
      required: [id, userId, pair, direction, timeframe, entryPrice, stopLoss, takeProfit, lotSize, riskPercent, rrRatio, reasoning, emotionScore, overallScore, scoreLabel, confluenceData, positives, warnings, negatives, verdict, devilsAdvocate, recommendations, waitFor, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        pair:
          type: string
        direction:
          type: string
        timeframe:
          type: string
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: number
        riskPercent:
          type: number
        rrRatio:
          type: number
        reasoning:
          type: string
        emotionScore:
          type: integer
        htfBias:
          type: ["string", "null"]
        overallScore:
          type: integer
        scoreLabel:
          type: string
        confluenceData:
          $ref: "#/components/schemas/Confluence"
        positives:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        negatives:
          type: array
          items:
            type: string
        verdict:
          type: string
        devilsAdvocate:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        waitFor:
          type: string
        newsAlert:
          oneOf:
            - $ref: "#/components/schemas/NewsAlert"
            - type: "null"
        propFirmAlert:
          type: ["string", "null"]
        revengeTradeWarning:
          type: boolean
        lotSizeRecommendation:
          type: ["number", "null"]
        screenshotUrl:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
    AnalysisSummary:
      type: object
      required: [id, pair, direction, overallScore, scoreLabel, outcome, createdAt]
      properties:
        id:
          type: string
        pair:
          type: string
        direction:
          type: string
        timeframe:
          type: string
        overallScore:
          type: integer
        scoreLabel:
          type: string
        verdict:
          type: string
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
    HistoryPage:
      type: object
      required: [analyses, total, page, limit]
      properties:
        analyses:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OutcomeUpdate:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [won, lost, cancelled, pending]
    HistoryFilter:
      type: object
      properties:
        pair:
          type: ["string", "null"]
        minScore:
          type: ["integer", "null"]
        maxScore:
          type: ["integer", "null"]
        outcome:
          type: ["string", "null"]
        page:
          type: ["integer", "null"]
        limit:
          type: ["integer", "null"]
    ChallengeInput:
      type: object
      required: [userId, firmName, accountSize, maxDrawdown, maxDailyLoss, profitTarget, startDate, phase]
      properties:
        userId:
          type: string
        firmName:
          type: string
        accountSize:
          type: number
        maxDrawdown:
          type: number
        maxDailyLoss:
          type: number
        profitTarget:
          type: number
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        phase:
          type: string
    PropChallenge:
      type: object
      required: [id, userId, firmName, accountSize, maxDrawdown, maxDailyLoss, profitTarget, startDate, phase, currentProfit, currentDrawdown, todayLoss, status, passingProbability, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        firmName:
          type: string
        accountSize:
          type: number
        maxDrawdown:
          type: number
        maxDailyLoss:
          type: number
        profitTarget:
          type: number
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        phase:
          type: string
        currentProfit:
          type: number
        currentDrawdown:
          type: number
        todayLoss:
          type: number
        status:
          type: string
        passingProbability:
          type: number
        daysRemaining:
          type: ["integer", "null"]
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/ChallengeTradeLog"
        createdAt:
          type: string
    ChallengeTradeLog:
      type: object
      required: [id, challengeId, pnl, createdAt]
      properties:
        id:
          type: string
        challengeId:
          type: string
        pnl:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ChallengeTradeInput:
      type: object
      required: [pnl]
      properties:
        pnl:
          type: number
        notes:
          type: ["string", "null"]
    SimulationInput:
      type: object
      required: [lossAmount]
      properties:
        lossAmount:
          type: number
    SimulationResult:
      type: object
      required: [newDrawdown, newDailyLoss, drawdownPercent, dailyLossPercent, wouldBreachDrawdown, wouldBreachDailyLoss, message]
      properties:
        newDrawdown:
          type: number
        newDailyLoss:
          type: number
        drawdownPercent:
          type: number
        dailyLossPercent:
          type: number
        wouldBreachDrawdown:
          type: boolean
        wouldBreachDailyLoss:
          type: boolean
        message:
          type: string
...
    TradeNote:
      type: object
      required: [id, analysisId, userId, note, createdAt]
      properties:
        id:
          type: string
        analysisId:
          type: string
        userId:
          type: string
        note:
          type: string
        mood:
          type: ["string", "null"]
        lessonsLearned:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TradeNoteInput:
      type: object
      required: [userId, note]
      properties:
        userId:
          type: string
        note:
          type: string
        mood:
          type: ["string", "null"]
        lessonsLearned:
          type: ["string", "null"]
...